requirements.txt
discord.py
yt-dlp
aiohttp
ffmpeg